Voice Leading
Implementation is fairly crude: bass note rises and falls to find the
chord root, other notes attempt to step down without dropping below
their lower neighbor. Does not prevent open fifths or parallel fifths or
anything.

There are many areas that are crude and hackish and require additional
tooling to be formalized. Also, I'm quite sure there are still bugs,
because the melody is behaving strangely and moving in ways it
shouldn't.

I will say, it's starting to sound a bit nice :)