Key Changes! Mostly. Major changes to PrettyProgressionStrategy, sundry
There is more work to be done. Clean up is first on the list, and I'm
sure there are bugs I haven't found yet. The melody is having some
trouble deciding what key to play in because it's offset from the
sections by one measure, and I haven't made it any smarter.